updating coverage keyword and restructuring
